// src/types/mockOverrides.d.ts
/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface MockOverrides {
  routes: {
    request: {
      path: string
      method: "get" | "post" | "put" | "delete"
    }
    responses: {
      statusCode: number
      headers?: {
        [k: string]: unknown
      }
      body?: {
        [k: string]: unknown
      }
    }[]
  }[]
}